pyproject.toml
[project]
name = "nabchan-mcp-server"
version = "0.1.0"
description = "Nablarchの解説書をもとにしてNablarchの情報を返すMCPサーバーです"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.13.3",
    "duckdb>=1.2.2",
    "html2text>=2024.2.26",
    "langchain-openai>=0.3.12",
    "lindera-py>=0.41.0",
    "mcp[cli]>=1.6.0",
    "pydantic>=2.11.2",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "sentencepiece>=0.2.0",
    "tiktoken>=0.9.0",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
]

[dependency-groups]
dev = [
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
lint = "uvx ruff format --check && uvx ruff check"
fix = "uvx ruff format && uvx ruff check --fix"
test = "DB_FILE=test.db python -m unittest discover -s tests"
integration_test = "DB_FILE=test.db python -m unittest discover -s tests -p it_*.py"
build_test_db = "DB_FILE=test.db python -m tests.db"
